% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_response_letter_codes.R
\name{add_response_letter_codes}
\alias{add_response_letter_codes}
\title{Add one letter codes response codes}
\usage{
add_response_letter_codes(dat_distill)
}
\arguments{
\item{dat_distill}{A tibble as output by \code{img_distill()} or \code{med_onc_distill()}}
}
\description{
Takes a distilled dataset and adds a column with single letter versions of the codes.  This is later useful for regex.
}
