% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_response_recist_style.R
\name{best_response_recist_style}
\alias{best_response_recist_style}
\title{Turn response codes (one letter) into the RECIST-style codes everyone is used to.}
\usage{
best_response_recist_style(resp_strings)
}
\description{
Turn response codes (one letter) into the RECIST-style codes everyone is used to.
}
