% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_unique_rows.R
\name{check_unique_rows}
\alias{check_unique_rows}
\title{Check uniqueness of rows in primary_keys}
\usage{
check_unique_rows(dat, primary_keys)
}
\arguments{
\item{dat}{A tibble/df that has the columns in primary_keys.}

\item{primary_keys}{A character vector containing the primary keys.}
}
\description{
Check uniqueness of rows in primary_keys
}
