% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_events.R
\name{filter_events}
\alias{filter_events}
\title{Filter img/med_onc events down to a range.}
\usage{
filter_events(
  df_time,
  df_time_key_cols,
  df_event,
  tol_low = 0.5,
  tol_high = 0.5
)
}
\arguments{
\item{df_time}{Dataframe with record_id, dob_event_min (minimum event time) and dob_event_max (maximum event time).}

\item{df_event}{An event dataframe of the style output by img_distill() or med_onc_distill().}

\item{tol_low}{Tolerance (days) surrounding dob_event_min.  The default of +0.5 will exclude events on dob_event_min.}

\item{tol_high}{Tolerance (days) surrounding dob_event_max.  The default of 0.5 will include events on dob_event_max.}
}
\description{
A simple join and subset, just a convenience function.
}
