% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_distill.R
\name{img_distill}
\alias{img_distill}
\title{Distill the imaging data}
\usage{
img_distill(dat_img, return_minimal = F)
}
\arguments{
\item{dat_img}{PRISSMM imaging dataset}

\item{return_minimal}{If true, returns fewer columns.}
}
\value{
A tibble with the same number of rows as dat_img.
}
\description{
Take the base BPC data and turn it into a format that can be more easily merged with med onc, other columns.
}
