% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_onc_distill.R
\name{med_onc_distill}
\alias{med_onc_distill}
\title{Distill the medical oncologist report data}
\usage{
med_onc_distill(dat_med_onc, return_minimal = F)
}
\arguments{
\item{dat_med_onc}{PRISSMM medical oncology dataset.}

\item{return_minimal}{If true, returns fewer columns.}
}
\value{
A tibble with the same number of rows as dat_img.
}
\description{
Take the base BPC data and turn it into a format that can be more easily merged with imaging, other columns.
}
