% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_sum_cs.R
\name{response_sum_cs}
\alias{response_sum_cs}
\title{Summarize the response to med onc OR imaging using only the cross sectional flags.}
\usage{
response_sum_cs(
  df_time,
  df_time_key_cols,
  df_event,
  time_tol_lb = 0.5,
  time_tol_ub = 0.5,
  event_type_override = NULL
)
}
\arguments{
\item{df_time}{A row of events to calculate responses to.  Must have columns record_id, dob_event_min, dob_event_max.}

\item{df_time_key_cols}{A character vector specifying the primary keys for this dataset (must be unique).}

\item{df_event}{An event dataframe of the style output by img_distill() or med_onc_distill().}
}
\description{
Cross sectional flags are those in status_change, or equivalently the one letter codes in resp_code_cs
}
\details{
Design note: This is a bulky function.  Time filtering has to be done in this function in addition to the summarization.  The reason is that the longitudinal versions (if ever written) will require baseline images to be fed in too, to establish where cancer has been found so far.
Confirmed response breakdown (the first one is a subset):
}
