% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_processor.R
\name{status_processor}
\alias{status_processor}
\title{Process the status column from med onc or imaging data.}
\usage{
status_processor(dat, col_name, output_name = "status_change")
}
\arguments{
\item{dat}{Any dataframe with a status column.}

\item{col_name}{String column name}
}
\description{
Takes the long names curators need like Progressing/Worsening/Enlarging and makes them into shorter things like "worsening"
}
